DiscreteEntropy

A Julia package for the estimation of Shannon entropy of discrete distributions.

Multiplicities

DiscreteEntropy uses the multiplicities representation of data. Given a histogram of samples

DiscreteEntropy.AbstractEstimator — Type

AbstractEstimator

Supertype for NonParameterised and Parameterised entropy estimators.

DiscreteEntropy.CountData — Type

CountData
an 2 x m matrix where m[1, :] is counts and m[2, :] the number of bins with those counts
[[2 3 1] => counts / icts
[2 1 2]] => bins / mm
so we have two bins with two, 1 bin with 3, and 2 bins with 1

DiscreteEntropy.EntropyData — Type

abstract type EntropyData

DiscreteEntropy.ansb — Method

ansb(data::CountData; undersampled::Float64=0.1)::Float64

Return the Asymptotic NSB estimation of the Shannon entropy of data in nats.

See Asymptotic NSB estimator (equations 11 and 12)

\[\hat{H}_{\tiny{ANSB}} = (C_\gamma - \log(2)) + 2 \log(N) - \psi(\Delta)\]

where $C_\gamma$ is Euler's Gamma ($\approx 0.57721...$), $\psi_0$ is the digamma function and $\Delta$ the number of coincidences in the data.

This is designed for the extremely undersampled regime (K ~ N) and diverges with N when well-sampled. ANSB requires that $N/K → 0$, which we set to be $N/K < 0.1$ by default

External Links

Asymptotic NSB estimator (equations 11 and 12)

DiscreteEntropy.bonachela — Method

bonachela(data::CountData)

Return the Bonachela estimator of the Shannon entropy of data in nats.

\[\hat{H}_{B} = \frac{1}{N+2} \sum_{i=1}^{K} \left( (h_i + 1) \sum_{j=n_i + 2}^{N+2} \frac{1}{j} \right)\]

External Links

Entropy estimates of small data sets

DiscreteEntropy.chao_shen — Method

chao_shen(data::CountData)

Return the Chao-Shen estimate of the Shannon entropy of data in nats.

\[\hat{H}_{CS} = - \sum_{i=i}^{K} \frac{\hat{p}_i^{CS} \log \hat{p}_i^{CS}}{1 - (1 - \hat{p}_i^{CS})}\]

where

\[\hat{p}_i^{CS} = (1 - \frac{1 - \hat{p}_i^{ML}}{N}) \hat{p}_i^{ML}\]

DiscreteEntropy.chao_wang_jost — Method

chao_wang_jost(data::CountData)

Return the Chao Wang Jost Shannon entropy estimate of data in nats.

\[\hat{H}_{\tiny{CWJ}} = \sum_{1 \leq h_i \leq N-1} \frac{h_i}{N} \left(\sum_{k=h_i}^{N-1} \frac{1}{k} \right) + \frac{f_1}{N} (1 - A)^{-N + 1} \left\{ - \log(A) - \sum_{r=1}^{N-1} \frac{1}{r} (1 - A)^r \right\}\]

with

\[A = \begin{cases} \frac{2 f_2}{(N-1) f_1 + 2 f_2} \, & \text{if} \, f_2 > 0 \\ \frac{2}{(N-1)(f_1 - 1) + 1} \, & \text{if} \, f_2 = 0, \; f_1 \neq 0 \\ 1, & \text{if} \, f_1 = f_2 = 0 \end{cases}\]

where $f_1$ is the number of singletons and $f_2$ the number of doubletons in data.

Notes

The algorithm is slightly modified port of that used in the entropart R library.

External Links

Entropy and the species accumulation curve: a novel entropy estimator via discovery rates of new species

DiscreteEntropy.conditional_entropy — Method

conditional_entropy(pmfX::AbstractVector{AbstractFloat}, pmfXY::AbstractVector{AbstractFloat})
conditional_entropy(X::CountData, XY::CountData, estimator::Type{T}) where {T<:NonParameterisedEstimator}
conditional_entropy(X::CountData, XY::CountData, estimator::Type{T}, param) where {T<:ParameterisedEstimator}

Compute the conditional entropy of Y conditioned on X

\[H(Y \mid X) = - \sum_{x \in X, y \in Y} p(x, y) \ln \frac{p(x, y)}{p(x)}\]

Compute the estimated conditional entropy of Y given X, from counts of X, and (X,Y) and estimator estimator

\[\hat{H}(Y \mid X) = \hat{H}(X, Y) - \hat{H}(X)\]

DiscreteEntropy.from_counts — Method

 from_counts(counts::CountVector, remove_zeros::Bool)

DiscreteEntropy.from_data — Method

from_data(data::AbstractVector, ::Type{Samples})
from_data(data::AbstractVector, ::Type{Histogram})

DiscreteEntropy.from_samples — Method

from_samples

DiscreteEntropy.grassberger — Method

grassberger(data::CountData)

Return the Grassberger (1988) estimation of Shannon entropy of data in nats

\[\hat{H}_{\tiny{Gr88}} = \sum_i \frac{h_i}{H} \left(\log(N) - \psi(h_i) - \frac{(-1)^{h_i}}{n_i + 1} \right)\]

Equation 13 from Finite sample corrections to entropy and dimension estimate

DiscreteEntropy.jackknife — Method

jackknife(data::CountData, statistic::Function; corrected=false)

Compute the jackknifed estimate of statistic on data.

DiscreteEntropy.jackknife_mle — Method

jackknife_mle(data::CountData; corrected=false)::Tuple{AbstractFloat, AbstractFloat}

Return the jackknifed estimate of data and the variance of the jackknifing (not the variance of the estimator itself).

If corrected is true, then the variance is scaled with n-1, else it is scaled with n

As found in the paper

DiscreteEntropy.jeffreys_divergence — Method

jeffreys_divergence(p, q)
(link)[https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7516653/]

\[J(p, q) = D_{KL}(p \Vert q) + D_{KL}(q \Vert p)\]

DiscreteEntropy.jensen_shannon_distance — Method

jensen_shannon_distance(P::AbstractVector, Q::AbstractVector, estimator)

Compute the Jensen Shannon Distance

DiscreteEntropy.jensen_shannon_divergence — Method

jenson_shannon_divergence(countsP::AbstractVector, countsQ::AbstractVector)
jenson_shannon_divergence(countsP::AbstractVector, countsQ::AbstractVector, estimator::Type{T}) where {T<:NonParamterisedEstimator}

Compute the Jenson Shannon Divergence between discrete distributions $P$ and $q$, as represented by their histograms.

\[\widehat{JS}(p, q) = \hat{H}\left(\frac{p + q}{2} \right) - \left( \frac{H(p) + H(q)}{2} \right) \]

DiscreteEntropy.kl_divergence — Method

kl_divergence(p::AbstractVector, q::AbstractVector)::Float64

\[D_{KL}(P ‖ Q) = \sum_{x \in X} P(x) \log \left( \frac{P(x)}{Q(x)} \right)\]

Compute the Kullback-Lebler Divergence between two discrete distributions. Both distributions needs to be defined over the same space, so length(p) == length(q). If the distributions are not normalised, they will be.

DiscreteEntropy.logx — Method

logx(x)::Float64

Returns natural logarithm of x, or 0.0 if x is zero

DiscreteEntropy.marginal_counts — Method

marginal_counts(contingency_matrix::Matrix, dim)

Return the unnormalised marginal counts of contingency_matrix along dimension dim.

DiscreteEntropy.maximum_likelihood — Method

maximum_likelihood(data::CountData)::Float64

Return the maximum likelihood estimation of Shannon entropy of data in nats.

\[\hat{H}_{\tiny{ML}} = - \sum_{i=1}^K p_i \log(p_i)\]

or equivalently

\[\hat{H}_{\tiny{ML}} = \log(N) - \frac{1}{N} \sum_{i=1}^{K}h_i \log(h_i)\]

DiscreteEntropy.miller_madow — Method

miller_madow(data::CountData)

Return the Miller Madow estimation of Shannon entropy, with a positive bias based on the total number of samples seen (N) and the support size (K).

\[\hat{H}_{\tiny{MM}} = \hat{H}_{\tiny{ML}} + \frac{K - 1}{2N}\]

DiscreteEntropy.nsb — Method

nsb(data, K=data.K)

Returns the Bayesian estimate of Shannon entropy of data, using the Nemenman, Shafee, Bialek algorithm

\[\hat{H}^{\text{NSB}} = \frac{ \int_0^{\ln(K)} d\xi \, \rho(\xi, \textbf{n}) \langle H^m \rangle_{\beta (\xi)} } { \int_0^{\ln(K)} d\xi \, \rho(\xi\mid n)}\]

where

\[\rho(\xi \mid \textbf{n}) = \mathcal{P}(\beta (\xi)) \frac{ \Gamma(\kappa(\xi))}{\Gamma(N + \kappa(\xi))} \prod_{i=1}^K \frac{\Gamma(n_i + \beta(\xi))}{\Gamma(\beta(\xi))}\]

DiscreteEntropy.pert — Method

pert(data::CountData, estimator)

A Pert estimate of entropy, where

H = \frac{a + 4b + c}{6}

where a is the minimum (maximum_likelihood), c is the maximum (log(k)) and $b$ is the most likely value (ChaoShen)

DiscreteEntropy.pym — Method

 pym(_mm::Vector{Int64}, _icts::Vector{Int64})

DiscreteEntropy.schurmann — Function

schurmann(data::CountData, ξ::Float64 = ℯ^(-1/2))

Return the Schurmann estimate of Shannon entropy of data in nats.

\[\hat{H}_{SHU} = \psi(N) - \frac{1}{N} \sum_{i=1}^{K} \, h_i \left( \psi(h_i) + (-1)^{h_i} ∫_0^{\frac{1}{\xi} - 1} \frac{t^{h_i}-1}{1+t}dt \right) \]

This is no one ideal value for $\xi$, however the paper suggests $e^{(-1/2)} \approx 0.6$

External Links

schurmann

DiscreteEntropy.schurmann_generalised — Method

schurmann_generalised(data::CountVector, xis::XiVector{T}) where {T<:Real}

schurmann_generalised

\[\hat{H}_{\tiny{SHU}} = \psi(N) - \frac{1}{N} \sum_{i=1}^{K} \, h_i \left( \psi(h_i) + (-1)^{h_i} ∫_0^{\frac{1}{\xi_i} - 1} \frac{t^{h_i}-1}{1+t}dt \right) \]

Return the generalised Schurmann entropy estimation, given a countvector data and a xivector xis, which must both be the same length.

schurmann_generalised(data::CountVector, xis::Distribution, scalar=false)

Computes the generalised Schurmann entropy estimation, given a countvector data and a distribution xis.

DiscreteEntropy.shrink — Method

shrink(data::CountData)

Return the Shrinkage, or James-Stein estimator of Shannon entropy for data in nats.

\[\hat{H}_{\tiny{SHR}} = - \sum_{i=1}^{K} \hat{p}_x^{\tiny{SHR}} \log(\hat{p}_x^{\tiny{SHR}})\]

where

\[\hat{p}_x^{\tiny{SHR}} = \lambda t_x + (1 - \lambda) \hat{p}_x^{\tiny{ML}}\]

and

\[\lambda = \frac{ 1 - \sum_{x=1}^{K} (\hat{p}_x^{\tiny{SHR}})^2}{(n-1) \sum_{x=1}^K (t_x - \hat{p}_x^{\tiny{ML}})^2}\]

with

\[t_x = 1 / K\]

Notes

Based on the implementation in the R package entropy

External Links

Entropy Inference and the James-Stein Estimator

DiscreteEntropy.to_bans — Method

to_bans(x::Float64)

Return $\frac{h}{log(10)}$ where h is in nats

DiscreteEntropy.to_bits — Method

to_bits(x::Float64)

Return $\frac{h}{\log(2)}$ where h is in nats

DiscreteEntropy.xlogx — Method

xlogx(x::Float64)

Returns x * log(x) for x ≥ 0, or 0.0 if x is zero

DiscreteEntropy.zhang — Method

zhang(data::CountData)

Return the Zhang estimate of the Shannon entropy of data in nats.

The recommended definition of Zhang's estimator is from Grabchak et al.

\[\hat{H}_Z = \sum_{i=1}^K \hat{p}_i \sum_{v=1}^{N - h_i} \frac{1}{v} ∏_{j=0}^{v-1} \left( 1 + \frac{1 - h_i}{N - 1 - j} \right)\]

The actual algorithm comes from Fast Calculation of entropy with Zhang's estimator by Lozano et al..

Links

Entropy estimation in turing's perspective