DiscreteEntropy

A Julia package for the estimation of Shannon entropy of discrete distributions.

Multiplicities

DiscreteEntropy uses the multiplicities representation of data. Given a histogram of samples

Types

DiscreteEntropy.AbstractEstimator — Type

AbstractEstimator

Supertype for NonParameterised and Parameterised entropy estimators.

DiscreteEntropy.estimate_h — Function

estimate_h(data::CountData, ::Type{T}) where {T<:AbstractEstimator}

Data

DiscreteEntropy.from_data — Function

from_data(data::AbstractVector, ::Type{Samples})
from_data(data::AbstractVector, ::Type{SampleHistogram})

DiscreteEntropy.from_samples — Function

from_samples

Missing docstring.

Missing docstring for from_counts. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EntropyData. Check Documenter's build log for details.

DiscreteEntropy.CountData — Type

CountData

Frequentist Estimators

DiscreteEntropy.maximum_likelihood — Function

maximum_likelihood(data::CountData)::Float64

Compute the maximum likelihood estimation of Shannon entropy of data in nats.

\[\hat{H}_{ML} = \log(N) - \frac{1}{N} \sum_{i=1}^{K}h_i \log(h_i)\]

Examples

julia> data = from_data([1,2,3,2,1], Histogram)
CountData([2.0 3.0 1.0; 2.0 1.0 2.0], 9.0, 6)

julia> maximum_likelihood(data)
1.522955067

Missing docstring.

Missing docstring for jackknife_ml. Check Documenter's build log for details.

DiscreteEntropy.miller_madow — Function

miller_madow(data::CountData)

Returns the maximum likelihood estimation of Shannon entropy, with a positive offset based on the total number of samples seen (N) and the support size (K).

\[\hat{H}_{MM} = \hat{H}_{ML} + \frac{K - 1}{2N}\]

DiscreteEntropy.grassberger — Function

grassberger(data::CountData)

Returns the Grassberger estimation of Shannon entropy.

\[\hat{H}_G = log(N) - \frac{1}{N} \sum_{i=1}^{K} h_i \; G(h_i)\]

This is essentially the same as $\hat{H}_{ML}$, but with the logarithm swapped for the scalar function $G$

where

\[G(h) = \psi(h) + \frac{1}{2}(-1)^h \big( \psi(\frac{h+1}{2} - \psi(\frac{h}{2}))\]

This is the solution to $G(h) = \psi(h) + (-1)^h \int_0^1 \frac{x^h - 1}{x+1} dx$ as given in the paper

DiscreteEntropy.schurmann — Function

schurmann(data::CountData, ξ::Float64 = ℯ^(-1/2))

schurmann

\[\hat{H}_{SHU} = \psi(N) - \frac{1}{N} \sum_{i=1}^{K} \, h_i \big( \psi(h_i) + (-1)^{h_i} ∫_0^{\frac{1}{\xi} - 1} \frac{t^{h_i}-1}{1+t}dt \big) \]

This is no one ideal value for $\xi$, however the paper suggests $e^{(-1/2)} \approx 0.6$

DiscreteEntropy.schurmann_generalised — Function

schurmann_generalised(data::CountVector, xis::XiVector{T}) where {T<:Real}

schurmann_generalised

\[\hat{H}_{SHU} = \psi(N) - \frac{1}{N} \sum_{i=1}^{K} \, h_i \big( \psi(h_i) + (-1)^{h_i} ∫_0^{\frac{1}{\xi_i} - 1} \frac{t^{h_i}-1}{1+t}dt \big) \]

Computes the generalised Schurmann entropy estimation, given a countvector data and a xivector xis, which must both be the same length.

schurmann_generalised(data::CountVector, xis::Distribution, scalar=false)

Computes the generalised Schurmann entropy estimation, given a countvector data and a distribution xis.

Missing docstring.

Missing docstring for zhang. Check Documenter's build log for details.

DiscreteEntropy.chao_shen — Function

chao_shen(data::CountData)

Return the Chao-Shen estimate of the the Shannon entropy of data in nats.

\[\hat{H}_{CS} = - \sum_{i=i}^{K} \frac{\hat{p}_i^{CS} \log \hat{p}_i^{CS}}{1 - (1 - \hat{p}_i^{CS})}\]

where

\[\hat{p}_i^{CS} = (1 - \frac{M}{N}) \hat{p}_i^{ML}\]

Missing docstring.

Missing docstring for bonachela. Check Documenter's build log for details.

Bayesian Estimators

Missing docstring.

Missing docstring for bayes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nsb. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ansb. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pym. Check Documenter's build log for details.

Resampling

We can also resample data

Missing docstring.

Missing docstring for jackknife. Check Documenter's build log for details.

Divergence

Missing docstring.

Missing docstring for kl_divergence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jeffreys_divergence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jensen_shannon_divergence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jensen_shannon_distance. Check Documenter's build log for details.

Conditional Entropy and Conditional Mutual Information

Conditional Entropy

Missing docstring.

Missing docstring for conditional_entropy. Check Documenter's build log for details.

Utilities

Missing docstring.

Missing docstring for logx. Check Documenter's build log for details.

Missing docstring.

Missing docstring for xlogx. Check Documenter's build log for details.

DiscreteEntropy.to_bits — Function

to_bits(x::Float64)

Return $\frac{h}{\log(2)}$ where h is in nats

DiscreteEntropy.to_bans — Function

to_bans(x::Float64)

Return $\frac{h}{log(10)}$ where h is in nats