DiscreteEntropy

A Julia package for the estimation of Shannon entropy of discrete distributions.

Multiplicities

DiscreteEntropy uses the multiplicities representation of data. Given a histogram of samples

Frequentist Estimators

DiscreteEntropy.maximum_likelihood — Function

maximum_likelihood(data::CountData)::Float64

Returns the maximum likelihood estimation of Shannon entropy.

\[\hat{H}_{ML} = \log(n) - \frac{1}{n} \sum_{k=1}^{K}h_k \log(h_k)\]

where n is the number of samples

DiscreteEntropy.miller_madow — Function

miller_madow(data::CountData)::Float64

Returns the maximum likelihood estimation of Shannon entropy, with a positive offset based on the total number of samples seen (n) and the support size (K).

\[\hat{H}_{MM} = \hat{H}_{ML} + \frac{K - 1}{2n}\]

DiscreteEntropy.grassberger — Function

grassberger(data::CountData)::Float64

Returns the Grassberger estimation of Shannon entropy.

\[\hat{H}_G = log(n) - \frac{1}{n} \sum_{k=1}^{K} h_k \; G(h_k)\]

This is essentially the same as $\hat{H}_{ML}$, but with the logarithm swapped for the scalar function $G$

where

\[G(h) = \psi(h) + \frac{1}{2}(-1)^h \big( \psi(\frac{h+1}{2} - \psi(\frac{h}{2}))\]

This is the solution to $G(h) = \psi(h) + (-1)^h \int_0^1 \frac{x^h - 1}{x+1} dx$ as given in the paper

DiscreteEntropy.schurmann — Function

schurmann(data::CountData, ξ::Float64 = ℯ^(-1/2))::Float64

schurmann

\[\hat{H}_{SHU} = \psi(n) - \frac{1}{n} \sum_{k=1}^{K} \, h_k \big( \psi(h_k) + (-1)^{h_k} ∫_0^{1/ζ - 1} \frac{t^{h_k}-1}{1+t}dt \big) \]

This is no one ideal value for $\zeta$, however the paper suggests $e^{(-1/2)} \approx 0.6$

DiscreteEntropy.zhang — Function

zhang(data::CountData)

Missing docstring.

Missing docstring for chao_shen. Check Documenter's build log for details.

DiscreteEntropy.bonachela — Function

bonachela(data::CountData)

Bayesian Estimators

DiscreteEntropy.bayes — Function

bayes(α::Float64, data::CountData)::Float64

Returns an estimate of Shannon entropy given data and a concentration parameter $α$.

\[\hat{H}_{\text{Bayes}} = - \sum_{k=1}^{K} \hat{p}_k^{\text{Bayes}} \; \log \hat{p}_k^{\text{Bayes}}\]

where

\[p_k^{\text{Bayes}} = \frac{k + α}{n + A}\]

and

\[A = \sum_{x=1}^{K} α_{x}\]

In addition to setting your own α, we have the following suggested choices

jeffrey : α = 0.5

laplace: α = 1.0

schurmann_grassberger: α = 1 / K

minimax: α = √{n} / K

DiscreteEntropy.nsb — Function

nsb(k, data::CountData)
nsb(data::CountData)
nsb(samples::AbstractVector)

Bayesian estimator

Utilities

DiscreteEntropy.xlogx — Function

xlogx(x::Float64)

Return x * log(x) for x ≥ 0, handling x == 0 by return 0.

DiscreteEntropy.to_bits — Function

to_bits(x::Float64)

Return $\frac{h}{\log(2)}$ where h is in nats

DiscreteEntropy.to_bans — Function

to_bans(x::Float64)

Return $\frac{h}{log(10)}$ where h is in nats