DiscreteEntropy

A Julia package for the estimation of Shannon entropy of discrete distributions.

Multiplicities

DiscreteEntropy uses the multiplicities representation of data. Given a histogram of samples

Types

DiscreteEntropy.AbstractEstimator — Type

AbstractEstimator

Supertype for NonParameterised and Parameterised entropy estimators.

DiscreteEntropy.estimate_h — Function

estimate_h(data::CountData, estimator::Type{T}) where {T<:AbstractEstimator}

Return the estimate in nats of Shannon entropy of data using estimator.

Data

DiscreteEntropy.from_data — Function

from_data(data::AbstractVector, ::Type{Samples})
from_data(data::AbstractVector, ::Type{SampleHistogram})

DiscreteEntropy.from_samples — Function

from_samples

Missing docstring.

Missing docstring for from_counts. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EntropyData. Check Documenter's build log for details.

DiscreteEntropy.CountData — Type

CountData

Frequentist Estimators

DiscreteEntropy.maximum_likelihood — Function

maximum_likelihood(data::CountData)::Float64

Compute the maximum likelihood estimation of Shannon entropy of data in nats.

\[\hat{H}_{ML} = \log(N) - \frac{1}{N} \sum_{i=1}^{K}h_i \log(h_i)\]

Examples

julia> data = from_data([1,2,3,2,1], Histogram)
CountData([2.0 3.0 1.0; 2.0 1.0 2.0], 9.0, 6)

julia> maximum_likelihood(data)
1.522955067

Missing docstring.

Missing docstring for jackknife_ml. Check Documenter's build log for details.

DiscreteEntropy.miller_madow — Function

miller_madow(data::CountData)

Returns the maximum likelihood estimation of Shannon entropy, with a positive offset based on the total number of samples seen (N) and the support size (K).

\[\hat{H}_{MM} = \hat{H}_{ML} + \frac{K - 1}{2N}\]

DiscreteEntropy.grassberger — Function

grassberger(data::CountData)

Returns the Grassberger estimation of Shannon entropy.

\[\hat{H}_G = log(N) - \frac{1}{N} \sum_{i=1}^{K} h_i \; G(h_i)\]

This is essentially the same as $\hat{H}_{ML}$, but with the logarithm swapped for the scalar function $G$

where

\[G(h) = \psi(h) + \frac{1}{2}(-1)^h \big( \psi(\frac{h+1}{2} - \psi(\frac{h}{2}))\]

This is the solution to $G(h) = \psi(h) + (-1)^h \int_0^1 \frac{x^h - 1}{x+1} dx$ as given in the paper

DiscreteEntropy.schurmann — Function

schurmann(data::CountData, ξ::Float64 = ℯ^(-1/2))

schurmann

\[\hat{H}_{SHU} = \psi(N) - \frac{1}{N} \sum_{i=1}^{K} \, h_i \big( \psi(h_i) + (-1)^{h_i} ∫_0^{\frac{1}{\xi} - 1} \frac{t^{h_i}-1}{1+t}dt \big) \]

This is no one ideal value for $\xi$, however the paper suggests $e^{(-1/2)} \approx 0.6$

DiscreteEntropy.schurmann_generalised — Function

schurmann_generalised(data::CountVector, xis::XiVector{T}) where {T<:Real}

schurmann_generalised

\[\hat{H}_{SHU} = \psi(N) - \frac{1}{N} \sum_{i=1}^{K} \, h_i \big( \psi(h_i) + (-1)^{h_i} ∫_0^{\frac{1}{\xi_i} - 1} \frac{t^{h_i}-1}{1+t}dt \big) \]

Computes the generalised Schurmann entropy estimation, given a countvector data and a xivector xis, which must both be the same length.

schurmann_generalised(data::CountVector, xis::Distribution, scalar=false)

Computes the generalised Schurmann entropy estimation, given a countvector data and a distribution xis.

DiscreteEntropy.zhang — Function

zhang(data::CountData)

Return the Zhang estimate of the Shannon entropy of data in nats.

The recommended definition of Zhang's estimator is from Grabchak et al.

\[\hat{H}_Z = \sum_{i=1}^K \hat{p}_i \sum_{v=1}^{N - h_i} \frac{1}{v} ∏_{j=0}^{v-1} \left( 1 + \frac{1 - h_i}{N - 1 - j} \right)\]

The actual algorithm comes from Fast Calculation of entropy with Zhang's estimator by Lozano et al..

Links

Entropy estimation in turing's perspective

DiscreteEntropy.chao_shen — Function

chao_shen(data::CountData)

Return the Chao-Shen estimate of the Shannon entropy of data in nats.

\[\hat{H}_{CS} = - \sum_{i=i}^{K} \frac{\hat{p}_i^{CS} \log \hat{p}_i^{CS}}{1 - (1 - \hat{p}_i^{CS})}\]

where

\[\hat{p}_i^{CS} = (1 - \frac{1 - \hat{p}_i^{ML}}{N}) \hat{p}_i^{ML}\]

Missing docstring.

Missing docstring for bonachela. Check Documenter's build log for details.

Bayesian Estimators

DiscreteEntropy.bayes — Function

bayes(data::CountData, α::AbstractFloat)

Returns an estimate of Shannon entropy given data and a concentration parameter $α$.

\[\hat{H}_{\text{Bayes}} = - \sum_{k=1}^{K} \hat{p}_k^{\text{Bayes}} \; \log \hat{p}_k^{\text{Bayes}}\]

where

\[p_k^{\text{Bayes}} = \frac{k + α}{n + A}\]

and

\[A = \sum_{x=1}^{K} α_{x}\]

In addition to setting your own α, we have the following suggested choices

jeffrey : α = 0.5

laplace: α = 1.0

schurmann_grassberger: α = 1 / K

minimax: α = √{n} / K

DiscreteEntropy.nsb — Function

nsb(data, K=data.K)

Returns the Bayesian estimate of Shannon entropy of data, using the Nemenman, Shafee, Bialek algorithm

\[\hat{H}^{\text{NSB}} = \frac{ \int_0^{\ln(K)} d\xi \, \rho(\xi, \textbf{n}) \langle H^m \rangle_{\beta (\xi)} } { \int_0^{\ln(K)} d\xi \, \rho(\xi\mid n)}\]

where

\[\rho(\xi \mid \textbf{n}) = \mathcal{P}(\beta (\xi)) \frac{ \Gamma(\kappa(\xi))}{\Gamma(N + \kappa(\xi))} \prod_{i=1}^K \frac{\Gamma(n_i + \beta(\xi))}{\Gamma(\beta(\xi))}\]

DiscreteEntropy.ansb — Function

ansb(data::CountData; undersampled::Float64=0.1)::Float64

\[\hat{H}_{ANSB} = \frac{C_\gamma}{\ln(2)} - 1 + 2 \ln(N) - \psi_0(\Delta)\]

where $C_\gamma$ is Euler's Gamma Constant $\approx 0.57721...$, $\psi_0$ is the digamma function and $\Delta$ the number of coincidences in the data.

Returns the Asymptotic NSB estimator (equations 11 and 12)

This is designed for the extremely undersampled regime (K ~ N) and diverges with N when well-sampled. ANSB requires that $N/K → 0$, which we set to be $N/K < 0.1$ by default

Missing docstring.

Missing docstring for pym. Check Documenter's build log for details.

Resampling

We can also resample data

Missing docstring.

Missing docstring for jackknife. Check Documenter's build log for details.

Divergence

Missing docstring.

Missing docstring for kl_divergence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jeffreys_divergence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jensen_shannon_divergence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jensen_shannon_distance. Check Documenter's build log for details.

Conditional Entropy and Conditional Mutual Information

Conditional Entropy

Missing docstring.

Missing docstring for conditional_entropy. Check Documenter's build log for details.

Other Useful Measures

DiscreteEntropy.redundancy — Function

redundancy(data::CountData, estimator::Type{T}) where {T<:AbstractEstimator}

Return the estimated information redundancy of data, with K the sampled support size.

\[R = \log(K) - \hat{H}(data)\]

redundancy(data::CountData, estimator::Type{T}, K::Int64) where {T<:AbstractEstimator}

Return the estimated information redundancy of data, with K set by the user.

External Links

(https://en.wikipedia.org/wiki/Redundancy(informationtheory))

DiscreteEntropy.uncertainty_coefficient — Function

uncertainty_coefficient(counts::Matrix)

\[C_{XY} = \frac{I(X;Y)}{H(Y)}\]

Utilities

DiscreteEntropy.logx — Function

logx(x)::Float64

Returns natural logarithm of x, or 0.0 if x is zero

DiscreteEntropy.xlogx — Function

xlogx(x::Float64)

Returns x * log(x) for x ≥ 0, or 0.0 if x is zero

DiscreteEntropy.to_bits — Function

to_bits(x::Float64)

Return $\frac{h}{\log(2)}$ where h is in nats

DiscreteEntropy.to_bans — Function

to_bans(x::Float64)

Return $\frac{h}{log(10)}$ where h is in nats