DiscreteEntropy

A Julia package for the estimation of Shannon entropy of discrete distributions.

Multiplicities

DiscreteEntropy uses the multiplicities representation of data. Given a histogram of samples

Frequentist Estimators

DiscreteEntropy.maximum_likelihood — Function

maximum_likelihood(data::CountData)::Float64

Returns the maximum likelihood estimation of Shannon entropy.

\[\hat{H}_{ML} = \log(n) - \frac{1}{n} \sum_{k=1}^{K}h_k \log(h_k)\]

where n is the number of samples

DiscreteEntropy.jackknife_ml — Function

jackknife_ml(data::CountData; corrected=false)::Tuple{Float64, Float64}

Returns the jackknifed estimate of data and the variance of the jackknifing (not the variance of the estimator itself).

If corrected in true, then the variance is scaled with n-1, else it is scaled with n

As found in the paper

DiscreteEntropy.miller_madow — Function

miller_madow(data::CountData)::Float64

Returns the maximum likelihood estimation of Shannon entropy, with a positive offset based on the total number of samples seen (n) and the support size (K).

\[\hat{H}_{MM} = \hat{H}_{ML} + \frac{K - 1}{2n}\]

DiscreteEntropy.grassberger — Function

grassberger(data::CountData)::Float64

Returns the Grassberger estimation of Shannon entropy.

\[\hat{H}_G = log(n) - \frac{1}{n} \sum_{k=1}^{K} h_k \; G(h_k)\]

This is essentially the same as $\hat{H}_{ML}$, but with the logarithm swapped for the scalar function $G$

where

\[G(h) = \psi(h) + \frac{1}{2}(-1)^h \big( \psi(\frac{h+1}{2} - \psi(\frac{h}{2}))\]

This is the solution to $G(h) = \psi(h) + (-1)^h \int_0^1 \frac{x^h - 1}{x+1} dx$ as given in the paper

DiscreteEntropy.schurmann — Function

schurmann(data::CountData, ξ::Float64 = ℯ^(-1/2))::Float64

schurmann

\[\hat{H}_{SHU} = \psi(n) - \frac{1}{n} \sum_{k=1}^{K} \, y_x \big( \psi(y_x) + (-1)^{y_x} ∫_0^{\frac{1}{\xi} - 1} \frac{t^{y_x}-1}{1+t}dt \big) \]

This is no one ideal value for $\xi$, however the paper suggests $e^{(-1/2)} \approx 0.6$

DiscreteEntropy.schurmann_generalised — Function

schurmann_generalised(data::CountData, xis::Vector{Float64})::Float64

schurmann_generalised

\[\hat{H}_{SHU} = \psi(n) - \frac{1}{n} \sum_{k=1}^{K} \, y_x \big( \psi(y_x) + (-1)^{y_x} ∫_0^{\frac{1}{\xi_x} - 1} \frac{t^{y_x}-1}{1+t}dt \big) \]

Accepts a vector is $ξ$ values, rather than just one.

DiscreteEntropy.zhang — Function

zhang(data::CountData)

DiscreteEntropy.chao_shen — Function

chao_shen(data::CountData)

wip

DiscreteEntropy.bonachela — Function

bonachela(data::CountData)

Bayesian Estimators

DiscreteEntropy.bayes — Function

bayes(α::Float64, data::CountData)::Float64

Returns an estimate of Shannon entropy given data and a concentration parameter $α$.

\[\hat{H}_{\text{Bayes}} = - \sum_{k=1}^{K} \hat{p}_k^{\text{Bayes}} \; \log \hat{p}_k^{\text{Bayes}}\]

where

\[p_k^{\text{Bayes}} = \frac{k + α}{n + A}\]

and

\[A = \sum_{x=1}^{K} α_{x}\]

In addition to setting your own α, we have the following suggested choices

jeffrey : α = 0.5

laplace: α = 1.0

schurmann_grassberger: α = 1 / K

minimax: α = √{n} / K

DiscreteEntropy.nsb — Function

nsb(data; k=data.K)

Returns the Bayesian estimate of Shannon entropy of data, using the Nemenman, Shafee, Bialek algorithm

\[\hat{H}^{\text{NSB}} = \frac{ \int_0^{\ln(K)} d\xi \, \rho(\xi, \textbf{n}) \langle H^m \rangle_{\beta (\xi)} } { \int_0^{\ln(K)} d\xi \, \rho(\xi\mid n)}\]

where

\[\rho(\xi \mid \textbf{n}) = \mathcal{P}(\beta (\xi)) \frac{ \Gamma(\kappa(\xi))}{\Gamma(N + \kappa(\xi))} \prod_{i=1}^K \frac{\Gamma(n_i + \beta(\xi))}{\Gamma(\beta(\xi))}\]

DiscreteEntropy.ansb — Function

ansb(data::CountData; undersampled::Float64=0.1)::Float64

\[\hat{H}_{ANSB} = \frac{C_\gamma}{\ln(2)} - 1 + 2 \ln(N) - \psi_0(\Delta)\]

where $C_\gamma$ is Euler's Gamma Constant $\approx 0.57721...$, $\psi_0$ is the digamma function and $\Delta$ the number of coincidences in the data.

Returns the Asymptotic NSB estimator (equations 11 and 12)

This is designed for the extremely undersampled regime (K ~ N) and diverges with N when well-sampled. ANSB requires that $N/K → 0$, which we set to be $N/K < 0.1$ by default

DiscreteEntropy.pym — Function

pym(mm::Vector{Int64}, icts::Vector{Int64})::Float64

A more or less faithful port of the original matlab code to Julia

Resampling

We can also resample data

DiscreteEntropy.jackknife — Function

jackknife(data::CountData, statistic::Function; corrected=false)

Returns the jacknifed estimate of statistic on data.

Divergence

DiscreteEntropy.kl_divergence — Function

kl_divergence(p::AbstractVector, q::AbstractVector)::Float64

\[D_{KL}(P ‖ Q) = \sum_{x \in X} P(x) \log \left( \frac{P(x)}{Q(x)} \right)\]

Returns the Kullback-Lebler Divergence between two discrete distributions. Both distributions needs to be defined over the same space, so length(p) == length(q). If the distributions are not normalised, they will be.

DiscreteEntropy.jeffreys_divergence — Function

jeffreys_divergence(p, q)
(link)[https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7516653/]

\[J(p, q) = D_{KL}(p \Vert q) + D_{KL}(q \Vert p)\]

DiscreteEntropy.jensen_shannon_divergence — Function

jenson_shannon_divergence(p::AbstractVector, q::AbstractVector)::Float64

Returns the Jenson Shannon Divergence between discrete distributions $p$ and $q$.

\[\widehat{JS}(p, q) = \hat{H}\left(\frac{p + q}{2} \right) - \left( \frac{H(p) + H(q)}{2} \right) \]

Utilities

DiscreteEntropy.logx — Function

logx(x)::Float64

Returns natural logarithm of x, or 0.0 if x is zero

DiscreteEntropy.xlogx — Function

xlogx(x::Float64)

Returns x * log(x) for x ≥ 0, or 0.0 if x is zero

DiscreteEntropy.to_bits — Function

to_bits(x::Float64)

Return $\frac{h}{\log(2)}$ where h is in nats

DiscreteEntropy.to_bans — Function

to_bans(x::Float64)

Return $\frac{h}{log(10)}$ where h is in nats